[juolly/issue-tracker #16] refactor: 바뀐 컴포넌트에 따라 수정
